{cmd,pkg}/certificates: service refactor (#2938)
