+++ pkg/certificate/authorization/doc.go
@@ -1,0 +1,8 @@
+// Copyright (C) 2019 Storj Labs, Inc.
+// See LICENSE for copying information.
+
+/*
+Package authorization is used for managing one-time-use certificate-signing-
+authorizations and claims.
+*/
+package authorization
