+++ pkg/certificate/certificateclient/doc.go
@@ -1,0 +1,7 @@
+// Copyright (C) 2019 Storj Labs, Inc.
+// See LICENSE for copying information.
+
+/*
+Package certificateclient contains the client for the certificate endpoint.
+*/
+package certificateclient
