+++ pkg/certificate/doc.go
@@ -1,0 +1,8 @@
+// Copyright (C) 2019 Storj Labs, Inc.
+// See LICENSE for copying information.
+
+/*
+Package certificate is responsible for managing certificate signing operations
+on peer identities' certificate chains.
+*/
+package certificate
